+++ fronts-client/src/components/FrontsEdit/Collection.tsx
@@ -5,7 +5,7 @@ import { AlsoOnDetail } from 'types/Collection';
 import { CardSets, Card as TCard } from 'types/Collection';
 import GroupDisplayComponent from 'components/GroupDisplay';
 import GroupLevel from 'components/clipboard/GroupLevel';
-import Card from './CollectionComponents/Card';
+import Card from '../card/Card';
 import CardLevel from 'components/clipboard/CardLevel';
 import { PosSpec, Move } from 'lib/dnd';
 import { Dispatch } from 'types/Store';
