+++ fronts-client/src/components/card/Card.tsx
@@ -25,13 +25,13 @@ import {
   editionsCardImageCriteria,
   DRAG_DATA_CARD_IMAGE_OVERRIDE,
 } from 'constants/image';
-import Sublinks from './Sublinks';
+import Sublinks from '../FrontsEdit/CollectionComponents/Sublinks';
 import {
   selectIsCardFormOpen,
   editorClearCardSelection,
 } from 'bundles/frontsUI';
 import { bindActionCreators } from 'redux';
-import ArticleMetaForm from '../ArticleMetaForm';
+import ArticleMetaForm from '../form/ArticleMetaForm';
 import { updateCardMetaWithPersist as updateCardMetaAction } from 'actions/Cards';
 import { EditMode } from 'types/EditMode';
 import { selectEditMode } from 'selectors/pathSelectors';
@@ -45,7 +45,7 @@ import DragIntentContainer from 'components/DragIntentContainer';
 import { CardTypes, CardTypesMap } from 'constants/cardTypes';
 import { RecipeCard } from 'components/card/recipe/RecipeCard';
 import { ChefCard } from 'components/card/chef/ChefCard';
-import { RecipeMetaForm } from '../RecipeMetaForm';
+import { RecipeMetaForm } from '../form/RecipeMetaForm';
 
 export const createCardId = (id: string) => `collection-item-${id}`;
