+++ fronts-client/src/components/form/FormContent.tsx
@@ -1,11 +1,10 @@
 import { styled } from 'constants/theme';
-
-export type FormSize = 'wide' | 'default';
+import { CardSizes } from 'types/Collection';
 
 export const FormContent = styled.div`
   flex: 3;
   display: flex;
-  flex-direction: ${(props: { size?: FormSize }) =>
+  flex-direction: ${(props: { size?: CardSizes }) =>
     props.size !== 'wide' ? 'column' : 'row'};
   margin-bottom: 40px;
 `;
